requirements.txt
hydra-core
omegaconf
Pillow
requests